+++ include/FileReader/FileManager.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <iostream>
+#include <vector>
+#include <algorithm>
+
+class FileManager
+{
+public:
+    std::string commentToken;
+
+    FileManager();
+    ~FileManager();
+
+    virtual void Init() = 0;
+    virtual void HandleData(std::string token, std::vector<std::string> splittedLine) = 0;
+    virtual void OnFileLoaded() = 0;
+};
