+++ include/FileReader/ObjFileManager.h
@@ -7,12 +7,6 @@
 
 #include "FileManager.h"
 
-#include "FileReader/Converter/Converter.h"
-#include "FileReader/Converter/GlmVec2Converter.h"
-#include "FileReader/Converter/GlmVec3Converter.h"
-#include "FileReader/Converter/StringConverter.h"
-#include "FileReader/Converter/ObjFaceConverter.h"
-
 #include "Renderer/Object.h"
 
 class ObjFileManager: public FileManager
@@ -22,13 +16,16 @@ private:
     std::vector<glm::vec3> normals;
     std::vector<glm::vec2> uv;
 
-    std::vector<std::array<int, 9>> faces;
-
     std::vector<Vertex> vertices;
 
     std::vector<std::shared_ptr<Mesh>> meshes;
 
     std::string currentMesh = "";
+
+    glm::vec2 GetGlmVec2(std::vector<std::string> splittedLine);
+    glm::vec3 GetGlmVec3(std::vector<std::string> splittedLine);
+    
+    void LoadVerticesFromFace(std::vector<std::string> splittedLine);
 public:
     const std::string commentToken = "#";
 
@@ -37,9 +34,6 @@ public:
     ObjFileManager();
 
     void Init();
-    void SaveData(std::string token, Converter* converter, std::vector<std::string> splittedLine);
-    void InterpretData();
-    void addVerticesTemp(std::vector<std::string> splittedLine);
-
-    Converter* GetConverterByToken(std::string token);
+    void HandleData(std::string token, std::vector<std::string> splittedLine);
+    void OnFileLoaded();
 };
