--- src/FileReader/Converter/Converter.cpp
@@ -1,10 +1,0 @@
-#include "FileReader/Converter/Converter.h"
-
-Converter::Converter(std::vector<std::string> tokens)
-    : tokens(tokens)
-{
-}
-
-Converter::~Converter()
-{
-}
