+++ src/FileReader/FileManager.cpp
@@ -1,0 +1,9 @@
+#include "FileReader/FileManager.h"
+
+FileManager::FileManager()
+{
+}
+
+FileManager::~FileManager()
+{
+}
